more phpstan strict checks and their consequences

+++ app/Stats/User.php
@@ -279,7 +279,7 @@ class User extends \Gazelle\BaseObject {
      * @return int number of unresolved reports
      */
     public function unresolvedReportsTotal(): int {
-        return self::$db->scalar("
+        return (int)self::$db->scalar("
             SELECT count(*)
             FROM reportsv2 AS r
             INNER JOIN torrents AS t ON (t.ID = r.TorrentID)
